move user to package
