+++ src/main/java/com/lenap/hermes/user/UserRecord.java
@@ -1,3 +1,3 @@
-package com.lenap.hermes;
+package com.lenap.hermes.user;
 
 public record UserRecord(Long id, String username, String email) {}
