+++ src/main/java/com/lenap/hermes/user/UserResource.java
@@ -1,4 +1,4 @@
-package com.lenap.hermes;
+package com.lenap.hermes.user;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
